--- a/pkg/broker/filter/stats_reporter.go
+++ b/pkg/broker/filter/stats_reporter.go
@@ -114,7 +114,7 @@ func register() {
 		&view.View{
 			Description: dispatchTimeInMsecM.Description(),
 			Measure:     dispatchTimeInMsecM,
-			Aggregation: view.Distribution(metrics.Buckets125(1, 100)...), // 1, 2, 5, 10, 20, 50, 100
+			Aggregation: view.Distribution(metrics.Buckets125(1, 10000)...), // 1, 2, 5, 10, 20, 50, 100, 1000, 5000, 10000
 			TagKeys:     []tag.Key{namespaceKey, triggerKey, brokerKey, triggerFilterTypeKey, triggerFilterSourceKey, responseCodeKey, responseCodeClassKey},
 		},
 		&view.View{