Fixes #1821 #1822 

## Proposed Changes
- A dashboard (Knative Eventing - Broker/Trigger)
- A dashboard (Knative Eventing - Source)
- Extend the broker ingress event processing latency bucket from 100ms to 10000ms (when creating processing latency graph, found most event processing latencies are located in buckets between 100ms to 10000ms)

![Screen Shot 2019-09-24 at 5 22 32 PM](https://user-images.githubusercontent.com/52978759/65619522-28315480-df75-11e9-8591-f30967aa4d22.png)
![Screen Shot 2019-09-24 at 5 22 41 PM](https://user-images.githubusercontent.com/52978759/65619626-54e56c00-df75-11e9-948b-91a030e88e8d.png)
![Screen Shot 2019-09-24 at 5 21 00 PM](https://user-images.githubusercontent.com/52978759/65619678-6a5a9600-df75-11e9-82e0-47aac94e8d17.png)

**Release Note**

<!-- Enter your extended release note in the below block. If the PR requires
additional action from users switching to the new release, include the string
"action required". If no release note is required, write "NONE". -->

```release-note

```