@nachocano 2019-09-25
This looks great!!! Thanks @grac3gao 

I changed my mind, can we do the same as you did for processing latencies buckets, to dispatch latencies bucket? Both in the broker ingress and broker filter handler?